更新 产品库存，下单数量，购物车数量支持小数，待 dev 同步代码

+++ fightGroupModule/src/main/java/com/github/chenlijia1111/fightGroup/common/requestVo/fightGroupProduct/FightGroupProductAddParams.java
@@ -8,6 +8,7 @@ import io.swagger.annotations.ApiModelProperty;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -61,7 +62,7 @@ public class FightGroupProductAddParams {
      */
     @ApiModelProperty("拼团商品参与总库存")
     @PropertyCheck(name = "拼团商品参与总库存", checkFunction = PositiveNumberCheck.class)
-    private Integer totalStockCount;
+    private BigDecimal totalStockCount;
 
     /**
      * 成团人数
