+++ fightGroupModule/src/main/java/com/github/chenlijia1111/fightGroup/entity/FightGroupProduct.java
@@ -7,6 +7,7 @@ import lombok.Getter;
 import lombok.Setter;
 import lombok.experimental.Accessors;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import javax.persistence.*;
 
@@ -77,7 +78,7 @@ public class FightGroupProduct {
     @ApiModelProperty("拼团商品参与总库存")
     @PropertyCheck(name = "拼团商品参与总库存")
     @Column(name = "total_stock_count")
-    private Integer totalStockCount;
+    private BigDecimal totalStockCount;
 
     /**
      * 拼团商品当前库存
@@ -85,7 +86,7 @@ public class FightGroupProduct {
     @ApiModelProperty("拼团商品当前库存")
     @PropertyCheck(name = "拼团商品当前库存")
     @Column(name = "stock_count")
-    private Integer stockCount;
+    private BigDecimal stockCount;
 
     /**
      * 成团人数
